Upload builds to web server + other improvements.
Upload build zips to host chrooted directory using sftp for the new
nightly page, instead of using IIS from the VM.

Make some style changes to the PowerShell code.

Write date and time before check/build to the log.

Only do builds for changes to files that are likely to change the
binaries.

Make the register-task.ps1 script nicer and run the task as the calling
user with a password read securely from the console.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>