Use non-overlay wx for arm64-windows* +dep scripts
Use the standard wxWidgets vcpkg port for Windows on ARM64 due to some
kind of build issue with the master overlay port on that configuration.

Add scripts to prime the vcpkg cache with necessary dependencies and
test the build for all 6 Windows architecture configurations.

Remove osslsigncode because we can use signtool now.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>