build-deps: fix setting build environment
Reorder the triplets and don't run the Visual Studio environment script
when the environment is the same. This is a workaround for the
environment script only working the first few times it is invoked.

Save and restore the environment when setting the build environment.

Do some minor refactoring and improvements.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>