Bring Windows nightlies back online, without ARM64
Fix Visual Studio CMake killing the script when run as a task by using
the CMake from WinGet instead.

Pass the path to the current `cl.exe` to CMake in the `CMAKE_C_COMPILER`
and `CMAKE_CXX_COMPILER` variables in case it has difficulties finding
it.

Remove `-hostarch` from the invocation of `Launch-VsDevShell.ps1`
because it does not support ARM64 hosts.

Disable the ARM64 build for now because these shits are breaking it.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>